Fix bug on HDF5Writer with external pmt
The commit that introduced the writing of the external PMT has a
bug. For a given file, if the external PMT is on, there is no problem,
the program will found the external PMT channel number and write it,
but if the external PMT is off, channel 0 will be missing in the hdf5 file.